+++ releases.md
@@ -1,5 +1,10 @@
 # Releases
 
+## Unreleased
+
+  - Add support for `pnpm` as well as `npm`.
+  - Add `bake lively:electron:install` that installs electron locally.
+
 ## v0.1.0
 
   - Initial implementation.